requirements.txt
google-generativeai==0.8.6
gradio==6.5.1
python-dotenv==1.2.1
imageio==2.37.2
requests==2.32.5
